+++ mono_repo/lib/src/commands/travis/travis_shell.dart
@@ -37,7 +37,7 @@ EXIT_CODE=0
 
 for PKG in \${PKGS}; do
   ${echoWithEvaluation(prettyAnsi, styleBold, r'PKG: ${PKG}')}
-  pushd "\${PKG}" || exit \$?
+  pushd "\${PKG}" > /dev/null || exit \$?
 
   PUB_EXIT_CODE=0
   pub $pubDependencyCommand --no-precompile || PUB_EXIT_CODE=\$?
@@ -45,7 +45,8 @@ for PKG in \${PKGS}; do
   if [[ \${PUB_EXIT_CODE} -ne 0 ]]; then
     EXIT_CODE=1
     ${safeEcho(prettyAnsi, red, "pub $pubDependencyCommand failed")}
-    popd
+    popd > /dev/null
+    echo
     continue
   fi
 
@@ -55,7 +56,8 @@ for PKG in \${PKGS}; do
 ${_shellCase('TASK', _calculateTaskEntries(commandsToKeys, prettyAnsi))}
   done
 
-  popd
+  popd > /dev/null
+  echo
 done
 
 exit \${EXIT_CODE}