+++ tool/travis.sh
@@ -38,7 +38,7 @@ EXIT_CODE=0
 
 for PKG in ${PKGS}; do
   echo -e "\033[1mPKG: ${PKG}\033[22m"
-  pushd "${PKG}" || exit $?
+  pushd "${PKG}" > /dev/null || exit $?
 
   PUB_EXIT_CODE=0
   pub upgrade --no-precompile || PUB_EXIT_CODE=$?
@@ -46,7 +46,8 @@ for PKG in ${PKGS}; do
   if [[ ${PUB_EXIT_CODE} -ne 0 ]]; then
     EXIT_CODE=1
     echo -e '\033[31mpub upgrade failed\033[0m'
-    popd
+    popd > /dev/null
+    echo
     continue
   fi
 
@@ -85,7 +86,8 @@ for PKG in ${PKGS}; do
     esac
   done
 
-  popd
+  popd > /dev/null
+  echo
 done
 
 exit ${EXIT_CODE}