+++ mono_repo/lib/src/commands/travis/travis_shell.dart
@@ -24,40 +24,68 @@ ${createdWith()}
 $windowsBoilerplate
 
 if [[ -z \${PKGS} ]]; then
-  ${safeEcho(prettyAnsi, red, "PKGS environment variable must be set!")}
-  exit 1
+  ${safeEcho(prettyAnsi, red, "PKGS environment variable must be set! - TERMINATING JOB")}
+  exit 64
 fi
 
 if [[ "\$#" == "0" ]]; then
-  ${safeEcho(prettyAnsi, red, "At least one task argument must be provided!")}
-  exit 1
+  ${safeEcho(prettyAnsi, red, "At least one task argument must be provided! - TERMINATING JOB")}
+  exit 64
 fi
 
-EXIT_CODE=0
+SUCCESS_COUNT=0
+declare -a FAILURES
 
 for PKG in \${PKGS}; do
   ${echoWithEvaluation(prettyAnsi, styleBold, r'PKG: ${PKG}')}
-  pushd "\${PKG}" >/dev/null || exit \$?
+  EXIT_CODE=0
+  pushd "\${PKG}" >/dev/null || EXIT_CODE=\$?
 
-  PUB_EXIT_CODE=0
-  pub $pubDependencyCommand --no-precompile || PUB_EXIT_CODE=\$?
+  if [[ \${EXIT_CODE} -ne 0 ]]; then
+    ${echoWithEvaluation(prettyAnsi, red, "PKG: '\${PKG}' does not exist - TERMINATING JOB")}
+    exit 64
+  fi
 
-  if [[ \${PUB_EXIT_CODE} -ne 0 ]]; then
-    EXIT_CODE=1
-    ${safeEcho(prettyAnsi, red, "pub $pubDependencyCommand failed")}
+  pub $pubDependencyCommand --no-precompile || EXIT_CODE=\$?
+
+  if [[ \${EXIT_CODE} -ne 0 ]]; then
+    ${echoWithEvaluation(prettyAnsi, red, "PKG: \${PKG}; 'pub $pubDependencyCommand' - FAILED  (\${EXIT_CODE})")}
+    FAILURES+=("\${PKG}; 'pub $pubDependencyCommand'")
   else
     for TASK in "\$@"; do
+      EXIT_CODE=0
       echo
       ${echoWithEvaluation(prettyAnsi, styleBold, r'PKG: ${PKG}; TASK: ${TASK}')}
 ${_shellCase('TASK', _calculateTaskEntries(commandsToKeys, prettyAnsi))}
+
+      if [[ \${EXIT_CODE} -ne 0 ]]; then
+        ${echoWithEvaluation(prettyAnsi, red, 'PKG: \${PKG}; TASK: \${TASK} - FAILED (\${EXIT_CODE})')}
+        FAILURES+=("\${PKG}; TASK: \${TASK}")
+      else
+        ${echoWithEvaluation(prettyAnsi, green, 'PKG: \${PKG}; TASK: \${TASK} - SUCCEEDED')}
+        SUCCESS_COUNT=\$((SUCCESS_COUNT + 1))
+      fi
+
     done
   fi
 
-  popd >/dev/null || exit \$?
+  echo
+  ${echoWithEvaluation(prettyAnsi, green, "SUCCESS COUNT: \${SUCCESS_COUNT}")}
+
+  if [ \${#FAILURES[@]} -ne 0 ]; then
+    ${echoWithEvaluation(prettyAnsi, red, 'FAILURES: \${#FAILURES[@]}')}
+    for i in "\${FAILURES[@]}"; do
+      ${echoWithEvaluation(prettyAnsi, red, "  \$i")}
+    done
+  fi
+
+  popd >/dev/null || exit 70
   echo
 done
 
-exit \${EXIT_CODE}
+if [ \${#FAILURES[@]} -ne 0 ]; then
+  exit 1
+fi
 ''';
 
 List<String> _calculateTaskEntries(
@@ -94,12 +122,14 @@ List<String> _calculateTaskEntries(
 
   taskEntries.sort();
 
-  final echoContent = wrapAnsi(
-    prettyAnsi,
-    red,
-    "Not expecting TASK '\${TASK}'. Error!",
-  );
-  addEntry('*', ['echo -e "$echoContent"', 'EXIT_CODE=1']);
+  addEntry('*', [
+    echoWithEvaluation(
+      prettyAnsi,
+      red,
+      "Unknown TASK '\${TASK}' - TERMINATING JOB",
+    ),
+    'exit 64',
+  ]);
   return taskEntries;
 }